Fix bug: image center
Image center:
The center of rotation of a 2D image of dimensions xdim x ydim is
defined by ((int)xdim/2, (int)(ydim/2)) (with the first pixel in the
upper left being (0,0). Note that for both xdim=ydim=65 and for
xdim=ydim=64, the center will be at (32,32). This is the same convention
as used in SPIDER and XMIPP. Origin offsets reported for individual
images translate the image to its center and are to be applied BEFORE
rotations.

IN MATLAB
Due to the first pixel in the upper left being (1,1), the center will be
at (33, 33) for both xdim=ydim=65 and for xdim=ydim=64